+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+## Fixed
+
+- Allow handover of (Form a MAT) projects with a TRN instead of incoming trust
+  UKPRN.
+
 ## [Release-110][release-110]
 
 ## Fixed